+++ plugins/NNModel/cpp/NNUGens.cpp
@@ -20,41 +20,6 @@ T* rtAlloc(World* world, size_t size=1) {
 
 namespace NN {
 
-struct NNID { int32 parentID; int32 nodeID; };
-inline bool operator<(const NNID& p1, const NNID& p2) {
-  return (p1.parentID != p2.parentID) ? 
-        p1.parentID < p2.parentID :
-        p1.nodeID < p2.nodeID;
-}
-
-class NNUGenRegister {
-public:
-  void add(int32 id, NN* ugen) {
-    Print("registering ugen: #(%d, %d) %p\n", ugen->mParent->mNode.mHash, id, ugen);
-    m_ugens[NNID{ugen->mParent->mNode.mHash, id}] = ugen;
-  }
-  void remove(int32 id, const NN* ugen) {
-    try {
-      m_ugens.erase(NNID{ugen->mParent->mNode.mHash, id});
-    } catch (const std::out_of_range&) { }
-  }
-  NN* get(int32 id, const Unit* unit) {
-    Print("getting ugen: #(%d, %d)\n", unit->mParent->mNode.mHash, id);
-    try {
-      return m_ugens.at(NNID{unit->mParent->mNode.mHash, id});
-    } catch (const std::out_of_range&) {
-      Print("NNUGen #(%d, %d) not found\n", unit->mParent->mNode.mHash, id);
-      return nullptr;
-    }
-  }
-private:
-  std::map<NNID, NN*> m_ugens;
-};
-
-// global UGen reference store, for setting and getting attributes
-static NNUGenRegister nnUGenRegister;
-
-
 static NNModelMethod* getModelMethod(NNModelDesc* model, float methodIdx) {
   if (model == nullptr) return nullptr;
 
@@ -68,13 +33,71 @@ void NN::clearOutputs(int nSamples) {
   ClearUnitOutputs(this, nSamples);
 }
 
+// ATTRIBUTES
+NNSetAttr::NNSetAttr(std::string name, int inputIdx, float initVal):
+    attrName(name), inputIdx(inputIdx), value(initVal), valUpdated(true) {}
+
+void NNSetAttr::update(Unit* unit, int nSamples) {
+  float newval = IN0(inputIdx);
+  if (newval != value) {
+    value = newval;
+    valUpdated = true;
+  }
+}
+
+// attributes are provided as additional input pairs (attrId, val) after model inputs
+void NN::setupAttributes() {
+  int i = UGenInputs::inputs + m_inDim;
+  while (i < numInputs()) {
+    int attrIdx = in0(i);
+    std::string attrName = m_modelDesc->getAttributeName(attrIdx, true);
+    if (!attrName.empty()) {
+      int inputIdx = i + 1;
+      NNSetAttr setter(attrName, inputIdx, in0(inputIdx));
+      m_attributes.push_back(setter);
+    } else {
+      Print("NNUGen: attribute #%d not found\n", attrIdx);
+    }
+    i += 2; // attrIdx, val
+  }
+}
+
+static void model_perform_attributes(NN* nn_instance) {
+  for(auto& a: nn_instance->m_attributes) {
+    if(!a.changed()) continue;
+    try {
+      nn_instance->m_model.set_attribute(a.attrName, {a.getStrValue()});
+      // print attr value if debugging
+      if (nn_instance->m_debug >= Debug::attributes) {
+        auto currVal = nn_instance->m_model.get_attribute_as_string(a.attrName);
+        Print("%s: %s\n", a.attrName.c_str(), currVal.c_str());
+      }
+    } catch (...) {
+      Print("NNUGen: can't set attribute %s\n", a.attrName.c_str());
+    }
+  };
+}
 
 // PERFORM
 
 void model_perform_load(NN* nn) {
-  Print("loading\n");
-  nn->m_model.load(nn->m_modelDesc->m_path);
-  Print("loaded\n");
+  auto path = nn->m_modelDesc->m_path;
+  if (nn->m_debug >= Debug::all)
+    Print("NNUGen: loading model %s\n", path.c_str());
+  int err = nn->m_model.load(nn->m_modelDesc->m_path);
+  if (err) {
+    Print("NNUGen: ERROR loading model %s\n", path.c_str());
+    nn->mDone = true;
+    return;
+  }
+  if(nn->m_warmup) {
+    if (nn->m_debug >= Debug::all)
+      Print("NNUGen: warming up model\n", path.c_str());
+    nn->warmupModel();
+  }
+  nn->setupAttributes();
+  if (nn->m_debug >= Debug::all)
+    Print("NNUGen: loaded %s\n", path.c_str());
 }
 
 #define MODEL_PERFORM_CUSTOM 0
@@ -106,9 +129,10 @@ void model_perform(NN* nn_instance) {
     in_model.push_back(&nn_instance->m_inModel[nn_instance->m_bufferSize * c]);
   for (int c(0); c < nn_instance->m_outDim; ++c)
     out_model.push_back(&nn_instance->m_outModel[nn_instance->m_bufferSize * c]);
-      nn_instance->m_model.perform(in_model, out_model,
-                                    nn_instance->m_bufferSize,
-                                    nn_instance->m_method->name, 1);
+  model_perform_attributes(nn_instance);
+  nn_instance->m_model.perform(in_model, out_model,
+                               nn_instance->m_bufferSize,
+                               nn_instance->m_method->name, 1);
 }
 void model_perform_loop(NN *nn_instance) {
   model_perform_load(nn_instance);
@@ -119,11 +143,13 @@ void model_perform_loop(NN *nn_instance) {
     out_model.push_back(&nn_instance->m_outModel[nn_instance->m_bufferSize * c]);
   while (!nn_instance->m_should_stop_perform_thread) {
     if (nn_instance->m_data_available_lock.try_acquire_for(
-            std::chrono::milliseconds(200))) {
-      if(!nn_instance->m_should_stop_perform_thread)
+      std::chrono::milliseconds(200))) {
+      if(!nn_instance->m_should_stop_perform_thread) {
+        model_perform_attributes(nn_instance);
         nn_instance->m_model.perform(in_model, out_model,
-                                    nn_instance->m_bufferSize,
-                                    nn_instance->m_method->name, 1);
+                                     nn_instance->m_bufferSize,
+                                     nn_instance->m_method->name, 1);
+      }
 
       nn_instance->m_result_available_lock.release();
     }
@@ -136,10 +162,13 @@ void model_perform_loop(NN *nn_instance) {
 
 void NN::next(int nSamples) {
 
-  if (!m_modelDesc->is_loaded() || mDone || !m_enabled) {
+  if (!m_modelDesc->is_loaded() || mDone) {
     ClearUnitOutputs(this, nSamples);
     return;
   };
+
+  // update attr setters
+  for (auto& a: m_attributes) a.update(this, nSamples);
 
   // copy inputs to circular buffer
   for (int c(0); c < m_inDim; ++c) {
@@ -179,12 +208,13 @@ NN::NN():
   m_compute_thread(nullptr), m_useThread(true),
   m_data_available_lock(0), m_result_available_lock(1),
   m_should_stop_perform_thread(false),
-  m_enabled(false),
   m_inBuffer(nullptr), m_outBuffer(nullptr),
   m_inModel(nullptr), m_outModel(nullptr)
 {
-  int32 m_ugenID = static_cast<int32>(in0(UGenInputs::ugenIdx));
-  nnUGenRegister.add(m_ugenID, this);
+
+  m_debug = static_cast<int>(in0(UGenInputs::debug));
+  m_warmup = in0(UGenInputs::warmup) > 0;
+
   auto modelIdx = static_cast<unsigned short>(in0(UGenInputs::modelIdx));
   m_modelDesc = gModels.get(modelIdx);
   if (m_modelDesc)
@@ -231,18 +261,16 @@ NN::NN():
   // don't use external thread on NRT
   if (!mWorld->mRealTime) m_useThread = false;
   if (m_useThread)
-    m_compute_thread = new std::thread(model_perform_loop, this);
+    m_compute_thread = std::make_unique<std::thread>(model_perform_loop, this);
   else
     model_perform_load(this);
 
   mCalcFunc = make_calc_function<NN, &NN::next>();
-  m_enabled = true;
   /* Print("NN: Ctor done\n"); */
 }
 
 NN::~NN() {
   /* Print("NN: Dtor\n"); */
-  nnUGenRegister.remove(m_ugenId, this);
   if (m_compute_thread) {
     // don't wait for join, it would stall the dsp chain
     // thread calls freeBuffers() when stopped
@@ -300,94 +328,15 @@ void NN::freeBuffers() {
   /* RTFree(mWorld, m_model); */
 }
 
+void NN::warmupModel() {
 
-// PARAMS
+  std::vector<float *> in_model, out_model;
+  for (int c(0); c < m_inDim; ++c)
+    in_model.push_back(&m_inModel[m_bufferSize * c]);
+  for (int c(0); c < m_outDim; ++c)
+    out_model.push_back(&m_outModel[m_bufferSize * c]);
 
-
-
-
-
-NNAttrUGen::NNAttrUGen():m_attrName() {
-
-  /* Print("NNATTR::CTor\n"); */
-  int32 ugenIdx = static_cast<int32>(in0(NNAttrInputs::modelIdx));
-  m_ugen = nnUGenRegister.get(ugenIdx, this);
-  /* Print("> ugen: #%d %p\n", ugenIdx, m_ugen); */
-  auto attrIdx = static_cast<unsigned short>(in0(NNAttrInputs::attrIdx));
-  if (m_ugen != nullptr)
-    m_attrName = m_ugen->m_modelDesc->getAttributeName(attrIdx);
-  /* Print("> attr #%d: %s\n", attrIdx, m_attrName.c_str()); */
-  if (m_attrName.empty()) {
-    mDone = true;
-    Unit* unit = this;
-    SETCALC(ClearUnitOutputs);
-    return;
-  }
-}
-
-
-
-NNSet::NNSet() {
-  set_calc_function<NNSet, &NNSet::next>();
-}
-static bool setAttribute(Backend& backend, const std::string& attrName, float value, bool warn=false) {
-/*   /1* Print("setting attr %s to %f\n", name.c_str(), value); *1/ */
-  std::vector<std::string> args = {std::to_string(value)};
-  try {
-    backend.set_attribute(attrName, args);
-  } catch (...) {
-    if (warn) Print("NNBackend: can't set attribute %s\n", attrName.c_str());
-    return false;
-  }
-  return true;
-}
-void NNSet::next(int nSamples) {
-  Unit* unit = this;
-  ClearUnitOutputs;
-  if (mDone) { return; };
-  float val = in0(UGenInputs::value);
-  if (!m_init || val != m_lastVal) {
-    setAttribute(m_ugen->m_model, m_attrName, val);
-    m_lastVal = val;
-    m_init = true;
-  }
-}
-
-NNGet::NNGet() {
-  set_calc_function<NNGet, &NNGet::next>();
-}
-static float getAttribute(Backend& backend, const std::string& name, bool warn) {
-  try {
-    auto value = backend.get_attribute(name)[0];
-    /* auto str = m_backend.get_attribute_as_string(name); */
-    /* Print("STR %s\n", str.c_str()); */
-    if (value.isInt()) {
-      return static_cast<float>(value.toInt());
-    }
-    else if (value.isBool()) {
-      return value.toBool() ? 1.0 : 0.0;
-    }
-    else if (value.isDouble()) {
-      Print("%s: %d\n", name.c_str(), value.toDouble());
-      return static_cast<float>(value.toDouble());
-    }
-    else {
-      if (warn) Print("NNGet: attribute '%s' has unsupported type.\n", name.c_str());
-      return 0;
-    }
-  } catch (...) {
-    if (warn) Print("NNGet: can't get attribute %s\n", name.c_str());
-    return 0;
-  }
-}
-
-void NNGet::next(int nSamples) {
-  Unit* unit = this;
-  ClearUnitOutputs;
-  if (mDone) {
-    return;
-  };
-  out0(0) = getAttribute(m_ugen->m_model, m_attrName, true);
+  m_model.perform(in_model, out_model, m_bufferSize, m_method->name, 1);
 }
 
 } // namespace NN
@@ -398,7 +347,5 @@ PluginLoad(NNUGens) {
   ft = inTable;
 
   registerUnit<NN::NN>(ft, "NNUGen", false);
-  registerUnit<NN::NNSet>(ft, "NNSet", false);
-  registerUnit<NN::NNGet>(ft, "NNGet", false);
   NN::Cmd::definePlugInCmds();
 }